Authorized updating, deleting functions

+++ src/components/twitter-card.tsx
@@ -54,9 +54,12 @@ import { toast } from "sonner";
 import { Textarea } from "./ui/textarea";
 import { Checkbox } from "./ui/checkbox";
 import { Label } from "./ui/label";
+import { useAuth } from "@/hooks/use-auth";
 const height = 100; // Chiều cao giới hạn cho 5 dòng nội dung
 
 export default function TweetCard(tweet: Tables<"tweets">) {
+	const { user } = useAuth();
+	const isAuthor = user?.id === tweet.author_id;
 	const [liked, setLiked] = useState(false);
 	const [retweeted, setRetweeted] = useState(false);
 	const [comments, setComments] = useState(Math.floor(Math.random() * 1000));
@@ -71,12 +74,12 @@ export default function TweetCard(tweet: Tables<"tweets">) {
 	const [data, setData] = useState<Tables<"tweets">>(tweet);
 
 	const handleChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-		setData({ ...data, content: e.target.value});
-	}
+		setData({ ...data, content: e.target.value });
+	};
 
 	const handleCheckboxChange = (checked: boolean) => {
 		setData({ ...data, tag: checked });
-	}
+	};
 
 	const [edit, setEdit] = useState(false);
 	const [deleted, setDeleted] = useState(false);
@@ -89,6 +92,13 @@ export default function TweetCard(tweet: Tables<"tweets">) {
 
 	const handleDelete = async (id: string) => {
 		// Xử lý xóa tweet ở đây
+		if (!isAuthor) {
+			toast.error("You are not authorized to delete this tweet.", {
+				position: "top-center",
+			});
+			return;
+		}
+
 		const res = await deleteTweet(id);
 		setDeleted(false); // Đóng dialog sau khi xóa
 
@@ -105,7 +115,14 @@ export default function TweetCard(tweet: Tables<"tweets">) {
 	};
 
 	const handleUpdate = async (id: string) => {
-		const res = await updateTweet(id, {...data,
+		if (!isAuthor) {
+			toast.error("You are not authorized to update this tweet.", {
+				position: "top-center",
+			});
+			return;
+		}
+		const res = await updateTweet(id, {
+			...data,
 			updated_at: new Date().toISOString(),
 		});
 		setEdit(false); // Đóng dialog sau khi cập nhật
@@ -119,7 +136,7 @@ export default function TweetCard(tweet: Tables<"tweets">) {
 				position: "top-center",
 			});
 		}
-	}
+	};
 
 	const handleLike = () => {
 		setLiked(!liked);
@@ -137,7 +154,7 @@ export default function TweetCard(tweet: Tables<"tweets">) {
 			const isOverflowing = contentRef.current.clientHeight > height; // Giới hạn 5 dòng (khoảng 100px mỗi dòng)
 			setNeedsTruncation(isOverflowing);
 		}
-	}, [isExpanded, data.content]);
+	}, [user, isExpanded, data.content]);
 
 	return (
 		<>
@@ -168,86 +185,106 @@ export default function TweetCard(tweet: Tables<"tweets">) {
 							</span>
 						</div>
 					</div>
-					<>
-						<DropdownMenu modal={false}>
-							<DropdownMenuTrigger asChild>
-								<Button variant="ghost" aria-label="Open menu" size="icon-lg">
-									<MoreHorizontalIcon />
-								</Button>
-							</DropdownMenuTrigger>
-							<DropdownMenuContent className="w-40" align="end">
-								<DropdownMenuLabel>Actions</DropdownMenuLabel>
-								<DropdownMenuGroup>
-									<DropdownMenuItem onSelect={() => setEdit(true)}>
-										Edit
-									</DropdownMenuItem>
-									<DropdownMenuItem onSelect={() => setDeleted(true)}>
-										Delete
-									</DropdownMenuItem>
-									<DropdownMenuItem disabled>Download</DropdownMenuItem>
-								</DropdownMenuGroup>
-							</DropdownMenuContent>
-						</DropdownMenu>
-						<Dialog open={edit} onOpenChange={setEdit}>
-							<DialogContent className="w-md sm:w-xl">
-								<DialogHeader>
-									<DialogTitle>Edit Tweet</DialogTitle>
-									<DialogDescription>
-										Make changes to your tweet content below.
-									</DialogDescription>
-								</DialogHeader>
-								<FieldSet>
-									<FieldGroup>
-										<Field>
-											<FieldLabel htmlFor="content">Content</FieldLabel>
-											<Textarea
-												id="content"
-												placeholder="Write your tweet content here."
-												rows={4}
-												className="resize-none max-h-80 overflow-y-auto"
-												value={data.content || ""}
-												onChange={handleChange}
-											/>
-											<div className="flex items-center gap-3">
-												<Checkbox id="tag" checked={Boolean(data.tag)} onCheckedChange={handleCheckboxChange} />
-												<Label htmlFor="tag">
-													This tweet contains sensitive content
-												</Label>
-											</div>
-										</Field>
-									</FieldGroup>
-								</FieldSet>
-								<DialogFooter>
-									<DialogClose asChild>
-										<Button variant="outline">Cancel</Button>
-									</DialogClose>
-									<Button variant="secondary" onClick={() => handleUpdate(tweet.id)}>Update</Button>
-								</DialogFooter>
-							</DialogContent>
-						</Dialog>
-						<Dialog open={deleted} onOpenChange={setDeleted}>
-							<DialogContent className="sm:max-w-[425px]">
-								<DialogHeader>
-									<DialogTitle>Delete Tweet</DialogTitle>
-									<DialogDescription>
-										Are you sure you want to delete this tweet? This action
-										cannot be undone.
-									</DialogDescription>
-								</DialogHeader>
-								<DialogFooter>
-									<DialogClose asChild>
-										<Button variant="outline">Cancel</Button>
-									</DialogClose>
-									<Button
-										variant="destructive"
-										onClick={() => handleDelete(tweet.id)}
-									>
-										Delete
+					{isAuthor ? (
+						<>
+							<DropdownMenu modal={false}>
+								<DropdownMenuTrigger asChild>
+									<Button variant="ghost" aria-label="Open menu" size="icon-lg">
+										<MoreHorizontalIcon />
 									</Button>
-								</DialogFooter>
-							</DialogContent>
-						</Dialog>
-					</>
+								</DropdownMenuTrigger>
+								<DropdownMenuContent className="w-40" align="end">
+									<DropdownMenuLabel>Actions</DropdownMenuLabel>
+									<DropdownMenuGroup>
+										<DropdownMenuItem onSelect={() => setEdit(true)} disabled={!isAuthor}>
+											Edit
+										</DropdownMenuItem>
+										<DropdownMenuItem onSelect={() => setDeleted(true)} disabled={!isAuthor}>
+											Delete
+										</DropdownMenuItem>
+										<DropdownMenuItem disabled>Download</DropdownMenuItem>
+									</DropdownMenuGroup>
+								</DropdownMenuContent>
+							</DropdownMenu>
+							<Dialog open={edit} onOpenChange={setEdit}>
+								<DialogContent className="w-md sm:w-xl">
+									<DialogHeader>
+										<DialogTitle>Edit Tweet</DialogTitle>
+										<DialogDescription>
+											Make changes to your tweet content below.
+										</DialogDescription>
+									</DialogHeader>
+									<FieldSet>
+										<FieldGroup>
+											<Field>
+												<FieldLabel htmlFor="content">Content</FieldLabel>
+												<Textarea
+													id="content"
+													placeholder="Write your tweet content here."
+													rows={4}
+													className="resize-none max-h-80 overflow-y-auto"
+													value={data.content || ""}
+													onChange={handleChange}
+												/>
+												<div className="flex items-center gap-3">
+													<Checkbox
+														id="tag"
+														checked={Boolean(data.tag)}
+														onCheckedChange={handleCheckboxChange}
+													/>
+													<Label htmlFor="tag">
+														This tweet contains sensitive content
+													</Label>
+												</div>
+											</Field>
+										</FieldGroup>
+									</FieldSet>
+									<DialogFooter>
+										<DialogClose asChild>
+											<Button variant="outline">Cancel</Button>
+										</DialogClose>
+										<Button
+											variant="secondary"
+											onClick={() => handleUpdate(tweet.id)}
+										>
+											Update
+										</Button>
+									</DialogFooter>
+								</DialogContent>
+							</Dialog>
+							<Dialog open={deleted} onOpenChange={setDeleted}>
+								<DialogContent className="sm:max-w-[425px]">
+									<DialogHeader>
+										<DialogTitle>Delete Tweet</DialogTitle>
+										<DialogDescription>
+											Are you sure you want to delete this tweet? This action
+											cannot be undone.
+										</DialogDescription>
+									</DialogHeader>
+									<DialogFooter>
+										<DialogClose asChild>
+											<Button variant="outline">Cancel</Button>
+										</DialogClose>
+										<Button
+											variant="destructive"
+											onClick={() => handleDelete(tweet.id)}
+										>
+											Delete
+										</Button>
+									</DialogFooter>
+								</DialogContent>
+							</Dialog>
+						</>
+					) : (
+						<svg
+							className="text-primary h-6 w-auto inline-block fill-current"
+							viewBox="0 0 24 24"
+						>
+							<g>
+								<path d="M23.643 4.937c-.835.37-1.732.62-2.675.733.962-.576 1.7-1.49 2.048-2.578-.9.534-1.897.922-2.958 1.13-.85-.904-2.06-1.47-3.4-1.47-2.572 0-4.658 2.086-4.658 4.66 0 .364.042.718.12 1.06-3.873-.195-7.304-2.05-9.602-4.868-.4.69-.63 1.49-.63 2.342 0 1.616.823 3.043 2.072 3.878-.764-.025-1.482-.234-2.11-.583v.06c0 2.257 1.605 4.14 3.737 4.568-.392.106-.803.162-1.227.162-.3 0-.593-.028-.877-.082.593 1.85 2.313 3.198 4.352 3.234-1.595 1.25-3.604 1.995-5.786 1.995-.376 0-.747-.022-1.112-.065 2.062 1.323 4.51 2.093 7.14 2.093 8.57 0 13.255-7.098 13.255-13.254 0-.2-.005-.402-.014-.602.91-.658 1.7-1.477 2.323-2.41z" />
+							</g>
+						</svg>
+					)}
 				</div>
 				<div
 					ref={contentRef}
